refactor: extract tab text function to own file
This is a prerequisite for using this function in another file.
